+++ R/q_functions.R
@@ -115,11 +115,17 @@
 #' need to set `test_args` correctly
 #'
 #' @return
+#' A named list of outputs.
 #' If `mode` is `power`, then a
-#' `power4test` object is returned.
+#' `power4test` object is set to the
+#' element `power4test`.
 #' If `mode` is `region`, then a
 #' `n_region_from_power` object is
-#' returned.
+#' set of the element `n_region_from_power`.
+#' If `mode` is `n_from_power`, then a
+#' `n_from_power` object is
+#' set to the `n_from_power` element.
+
 #'
 #' @inheritParams power4test
 #' @inheritParams n_region_from_power
@@ -149,6 +155,14 @@
 #' @param parallel If `TRUE`, parallel
 #' processing will be used when calling
 #' other functions, if appropriate.
+#'
+#' @param algorithm The algorithm to be
+#' used in mode `"region"` and `"n"`.
+#' If `NULL`, then it will be determined
+#' internally based on the `mode`. (The
+#' default may be different from that
+#' of [n_region_from_power()] and
+#' [n_from_power()])
 #'
 #' @param ... For `q_power_mediation_*`,
 #' these are optional arguments to
@@ -248,10 +262,11 @@ q_power_mediation <- function(
   iseed = NULL,
   parallel = TRUE,
   progress = TRUE,
-  simulation_progress = TRUE,
-  max_trials = 10,
+  simulation_progress = NULL,
+  max_trials = NULL,
+  algorithm = NULL,
   ...,
-  mode = c("power", "region")
+  mode = c("power", "region", "n")
 ) {
   # A general quick functions for other
   # wrappers.
@@ -266,6 +281,10 @@ q_power_mediation <- function(
                             "will be estimated."),
               region = paste("The region of sample sizes",
                              "for a power of",
+                              sprintf("%3.2f", target_power),
+                             "will be searched."),
+              n = paste("A sample size",
+                             "with estimated power close to",
                               sprintf("%3.2f", target_power),
                              "will be searched.")
     )
@@ -304,6 +323,9 @@ q_power_mediation <- function(
 
   n_region_from_power_args <- names(formals(n_region_from_power))
   ddd_n_from_region <- ddd[names(ddd) %in% n_region_from_power_args]
+
+  n_from_power_args <- names(formals(n_from_power))
+  ddd_n_from_power <- ddd[names(ddd) %in% n_from_power_args]
 
   # No need. This is the new default behavior of n_from_region
   # if (is.null(ddd_n_from_region$final_nrep)) {
@@ -354,6 +376,11 @@ q_power_mediation <- function(
 
   if (mode == "region") {
 
+    if (is.null(algorithm)) {
+      # Do not override the default, for now
+      algorithm <- NULL
+    }
+
     # Do this only if mode is not "power",
     # i.e., "region"
 
@@ -363,6 +390,7 @@ q_power_mediation <- function(
               progress = progress,
               simulation_progress = simulation_progress,
               max_trials = max_trials,
+              algorithm = algorithm,
               seed = seed
             )
 
@@ -380,8 +408,40 @@ q_power_mediation <- function(
     out2 <- NULL
   }
 
+  # ==== Find the n ====
+
+  if (mode == "n") {
+
+    if (is.null(algorithm)) {
+      # Do not override the default, for now
+      algorithm <- "probabilistic_bisection"
+    }
+
+    n_args <- list(
+              object = out,
+              target_power = target_power,
+              progress = progress,
+              simulation_progress = simulation_progress,
+              max_trials = max_trials,
+              algorithm = algorithm,
+              seed = seed
+            )
+
+    n_args <- utils::modifyList(
+                  n_args,
+                  ddd_n_from_power
+                )
+    out3 <- do.call(
+              n_from_power,
+              n_args
+            )
+  } else {
+    out3 <- NULL
+  }
+
   outz <- list(power4test = out,
-               n_region_from_power = out2)
+               n_region_from_power = out2,
+               n_from_power = out3)
   class(outz) <- c("q_power_mediation",
                    class(outz))
   outz
@@ -393,10 +453,14 @@ q_power_mediation <- function(
 #' @param mode What to print. If `"region"`
 #' and the output of [n_region_from_power()]
 #' is available, it will print the results
-#' of [n_region_from_power()].
+#' of [n_region_from_power()] if available.
+#' If `"n"`, then the output of
+#' [n_from_power()] will be printed if
+#' available.
 #' If `"power"`, then the output of
-#' [power4test()] will be printed. If
-#' `"all"`, then all available output
+#' [power4test()] will be printed.
+#' If
+#' `"all"`, then all available outputs
 #' will be printed.
 #'
 #' @return
@@ -408,7 +472,7 @@ q_power_mediation <- function(
 #' @export
 print.q_power_mediation <- function(
   x,
-  mode = c("all", "region", "power"),
+  mode = c("all", "region", "n", "power"),
   ...
 ) {
   mode <- match.arg(mode)
@@ -428,11 +492,24 @@ print.q_power_mediation <- function(
   # ==== Print n_region_from_power ====
 
   if (mode %in% c("region", "all")) {
-    cat("\n========== n_region_from_power Results ==========\n\n")
     if (is.null(x$n_region_from_power)) {
-      cat("\n'mode' is not 'region' and results not available.\n\n")
+      # cat("\n'mode' is not 'region' and results not available.\n\n")
     } else {
+      cat("\n========== n_region_from_power Results ==========\n\n")
       print(x$n_region_from_power,
+            ...)
+    }
+  }
+
+  # ==== Print n_from_power ====
+
+  if (mode %in% c("n", "all")) {
+    if (is.null(x$n_from_power)) {
+      # cat("\n'mode' is not 'n_from_power' and results not available.\n\n")
+    } else {
+      cat("\n========== n_from_power Results ==========\n\n")
+      print(x$n_from_power,
+            call = FALSE,
             ...)
     }
   }
@@ -442,8 +519,14 @@ print.q_power_mediation <- function(
 
 #' @return
 #' The `plot`-method of `q_power_mediation`
-#' returns `x` invisibly.
-#' It is called for its side effect.
+#' returns `NULL`.
+#' It will plot either the output of
+#' [n_region_from_power()]
+#' (mode `"region"`) or the output of
+#' [n_from_power()] (mode `"n"`).
+#' If no output from
+#' either of these functions is available,
+#' nothing wil be plotted.
 #'
 #' @rdname q_power_mediation
 #' @export
@@ -451,11 +534,16 @@ plot.q_power_mediation <- function(
   x,
   ...
 ) {
-  if (is.null(x$n_region_from_power)) {
-    stop("'mode' is not 'region' and nothing to plot.")
+  # x will have only one of the followings
+  if (!is.null(x$n_region_from_power)) {
+    plot(x$n_region_from_power,
+         ...)
   }
-  plot(x$n_region_from_power,
-       ...)
+  if (!is.null(x$n_from_power)) {
+    plot(x$n_from_power,
+         ...)
+  }
+  NULL
 }
 
 #' @param object For the `summary`
@@ -465,22 +553,31 @@ plot.q_power_mediation <- function(
 #' @return
 #' The `summary` method for
 #' `q_power_mediation` objects returns
-#' the output of [summary.n_region_from_power()].
-#' An error is raised if the output
-#' of [n_region_from_power()] is not
-#' available.
+#' the output of [summary()] for
+#' either the output of [n_region_from_power()]
+#' (mode `"region"`) or the output of
+#' [n_from_power()] (mode `"n"`).
+#' Return `NULL` if no output from
+#' either of these functions is available.
 #'
 #' @rdname q_power_mediation
 #' @export
 summary.q_power_mediation <- function(
                                 object,
                                 ...) {
-  if (is.null(object$n_region_from_power)) {
-    stop("'mode' is not 'region' and summary cannot be generated.")
+  if (!is.null(object$n_region_from_power)) {
+    # stop("'mode' is not 'region' and summary cannot be generated.")
+    out <- summary(object$n_region_from_power,
+                   ...)
+    return(out)
   }
-  out <- summary(object$n_region_from_power,
-                 ...)
-  out
+  if (!is.null(object$n_from_power)) {
+    # stop("'mode' is not 'region' and summary cannot be generated.")
+    out <- summary(object$n_from_power,
+                   ...)
+    return(out)
+  }
+  NULL
 }
 
 #' @param a For a simple mediation
@@ -541,10 +638,11 @@ q_power_mediation_simple <- function(
   iseed = NULL,
   parallel = TRUE,
   progress = TRUE,
-  simulation_progress = TRUE,
-  max_trials = 10,
+  simulation_progress = NULL,
+  max_trials = NULL,
+  algorithm = NULL,
   ...,
-  mode = c("power", "region")
+  mode = c("power", "region", "n")
 ) {
 
   # ==== Setup the model =====
@@ -626,6 +724,7 @@ q_power_mediation_simple <- function(
     progress = progress,
     simulation_progress = simulation_progress,
     max_trials = max_trials,
+    algorithm = algorithm,
     ...,
     mode = mode
   )
@@ -690,10 +789,11 @@ q_power_mediation_serial <- function(
   iseed = NULL,
   parallel = TRUE,
   progress = TRUE,
-  simulation_progress = TRUE,
-  max_trials = 10,
+  simulation_progress = NULL,
+  max_trials = NULL,
+  algorithm = NULL,
   ...,
-  mode = c("power", "region")
+  mode = c("power", "region", "n")
 ) {
 
   # ==== Setup the model =====
@@ -817,6 +917,7 @@ q_power_mediation_serial <- function(
     progress = progress,
     simulation_progress = simulation_progress,
     max_trials = max_trials,
+    algorithm = algorithm,
     ...,
     mode = mode
   )
@@ -908,10 +1009,11 @@ q_power_mediation_parallel <- function(
   iseed = NULL,
   parallel = TRUE,
   progress = TRUE,
-  simulation_progress = TRUE,
-  max_trials = 10,
+  simulation_progress = NULL,
+  max_trials = NULL,
+  algorithm = NULL,
   ...,
-  mode = c("power", "region")
+  mode = c("power", "region", "n")
 ) {
 
   omnibus <- match.arg(omnibus)
@@ -1016,6 +1118,7 @@ q_power_mediation_parallel <- function(
     progress = progress,
     simulation_progress = simulation_progress,
     max_trials = max_trials,
+    algorithm = algorithm,
     ...,
     mode = mode
   )