+++ R/x_from_power_algorithm_probabilistic_bisection.R
@@ -282,7 +282,9 @@ power_algorithm_prob_bisection <- function(
                                           n = TRUE,
                                           NULL),
                     rollback = TRUE,
-                    bz = TRUE)
+                    bz = TRUE,
+                    min_interval_width = c(n = 100,
+                                           es = .20))
   variants <- utils::modifyList(variants0,
                                 variants)
   if (is.null(variants$hdr_prob)) {
@@ -685,7 +687,7 @@ power_algorithm_prob_bisection <- function(
                                         R = R,
                                         what = f_what,
                                         goal = f_goal,
-                                        tol = tol,
+                                        tol = -Inf,
                                         simulation_progress = simulation_progress,
                                         save_sim_all = save_sim_all,
                                         progress = extend_progress,
@@ -703,10 +705,12 @@ power_algorithm_prob_bisection <- function(
                                         digits = digits,
                                         store_output = TRUE,
                                         overshoot = switch(x,
-                                                           n = .5,
-                                                           es = .05),
+                                                           n = 1,
+                                                           es = .10),
                                         variants = variants,
-                                        proxy_power = proxy_power)
+                                        proxy_power = proxy_power,
+                                        extend_only = TRUE
+                                        )
     by_x_1 <- interval_updated$by_x_1 %||% by_x_1
     ci_hit <- FALSE
     solution_found <- FALSE
@@ -830,7 +834,7 @@ power_algorithm_prob_bisection <- function(
     ok_lower <- tmp["ok_lower"]
     ok_upper <- tmp["ok_upper"]
 
-    if ((interval_updated$extend_status != 0) &&
+    if ((interval_updated$extend_status > 0) &&
         (!ok_lower && !ok_upper)) {
 
       # ==== No solution and interval invalid. Skip the search ====
@@ -1038,8 +1042,10 @@ power_algorithm_prob_bisection <- function(
         # ==== New output ====
         if (do_final_check) {
           tmp_R <- final_R
+          tmp_sim_progress <- progress
         } else {
           tmp_R <- R
+          tmp_sim_progress <- simulation_progress
         }
         out_i <- f(
                   x_i = x_i,
@@ -1053,7 +1059,7 @@ power_algorithm_prob_bisection <- function(
                   nrep = nrep_i,
                   R = tmp_R,
                   what = f_what,
-                  simulation_progress = simulation_progress,
+                  simulation_progress = tmp_sim_progress,
                   save_sim_all = save_sim_all,
                   store_output = TRUE,
                   target_nrep = final_nrep,
@@ -1345,9 +1351,11 @@ power_algorithm_prob_bisection <- function(
         hdr_power_pb <- hdr_power_pb / sum(hdr_power_pb)
         hdr_power_dominant <- hdr_power_pb > .80
         hdr_power_dominant_i <- which(hdr_power_dominant)
-      } else {
+      } else if (length(hdr_power) == 1) {
         hdr_power_dominant <- TRUE
         hdr_power_dominant_i <- 1
+      } else {
+        hdr_power_dominant <- FALSE
       }
       terminate_hdr_power_tol <- FALSE
       if (any(hdr_power_dominant) &&
@@ -1376,32 +1384,61 @@ power_algorithm_prob_bisection <- function(
 
       # ==== Termination: last_k changes ====
 
+      terminate_change_ok_x <- FALSE
+      terminate_change_ok_f <- FALSE
+
       # Do this check even if doing a final check
-
-      if ((!changes_ok || !changes_ok_f) &&
-          (perturbation_count == 0)) {
-
-        status <- bisection_status_message(3, status)
-
+      # Not much change
+      # - A final check but not a solution? -> Perturbation
+      # - A final check and solution found? -> change_ok ignored
+      # - A PBA iteration -> Do a final check next
+      if (do_final_check) {
+        if (!ok) {
+          if (!changes_ok || !changes_ok_f) {
+            if (perturbation_count > 0) {
+              perturbation_count <- perturbation_count - 1
+              x_i <- x_i + delta_tol * 50 * z_i
+              x_i <- min(max(lower_i, x_i), upper_i)
+              x_i <- switch(
+                        x_type,
+                        n = ceiling(x_i),
+                        es = x_i
+                      )
+            } else {
+              status <- bisection_status_message(3, status)
+              terminate_change_ok_x <- !changes_ok
+              terminate_change_ok_f <- !changes_ok_f
+            }
+          }
+        }
+      } else {
         terminate_change_ok_x <- !changes_ok
         terminate_change_ok_f <- !changes_ok_f
-
-      } else if ((!changes_ok || !changes_ok_f) &&
-                 !ok &&
-                 do_final_check) {
-        # If not a solution, then the next iteration is a PBA iteration
-        perturbation_count <- perturbation_count - 1
-        x_i <- x_i + delta_tol * 50 * z_i
-        x_i <- min(max(lower_i, x_i), upper_i)
-        # x_i <- q_dfun(dfun_i,
-        #               prob = runif(1, -.10, .10) +
-        #                      sample(c(.30, .70), size = 1))
-        x_i <- switch(
-                  x_type,
-                  n = ceiling(x_i),
-                  es = x_i
-                )
       }
+      # if ((!changes_ok || !changes_ok_f) &&
+      #     (perturbation_count == 0)) {
+
+      #   status <- bisection_status_message(3, status)
+
+      #   terminate_change_ok_x <- !changes_ok
+      #   terminate_change_ok_f <- !changes_ok_f
+
+      # } else if ((!changes_ok || !changes_ok_f) &&
+      #            !ok &&
+      #            do_final_check) {
+      #   # If not a solution, then the next iteration is a PBA iteration
+      #   perturbation_count <- perturbation_count - 1
+      #   x_i <- x_i + delta_tol * 50 * z_i
+      #   x_i <- min(max(lower_i, x_i), upper_i)
+      #   # x_i <- q_dfun(dfun_i,
+      #   #               prob = runif(1, -.10, .10) +
+      #   #                      sample(c(.30, .70), size = 1))
+      #   x_i <- switch(
+      #             x_type,
+      #             n = ceiling(x_i),
+      #             es = x_i
+      #           )
+      # }
 
       # ==== Next i: Set nrep ====
 
@@ -1963,68 +2000,74 @@ power_algorithm_prob_bisection_pre_i <- function(object,
 
   progress_type <- match.arg(progress_type)
 
-  # ==== Initial values ====
+  # PBA does not extend the range dynamically (for now).
+  # Moreover, the initial nrep is usually very small.
+  # Therefore, use x_interval as-is. Do not try a narrower interval.
 
-  # This method only needs an initial interval
-  if (inherits(object_by_org, "power4test_by_n") ||
-      inherits(object_by_org, "power4test_by_es")) {
-    x_i <- set_x_range_by_x(
-                      object,
-                      x = x,
-                      pop_es_name = pop_es_name,
-                      target_power = target_power,
-                      k = 2,
-                      x_max = x_max,
-                      x_min = x_min,
-                      object_by_org = object_by_org,
-                      what = what,
-                      goal = goal,
-                      tol = tol,
-                      ci_level = ci_level)
-  } else {
-    x_i <- set_x_range(object,
-                      x = x,
-                      pop_es_name = pop_es_name,
-                      target_power = target_power,
-                      k = 2,
-                      x_max = x_max,
-                      x_min = x_min)
-  }
+  x_i <- x_interval
 
-  # For bisection, no need to exclude the value in the input objects
-  # Exclude the value in the input object
+  # # ==== Initial values ====
 
-  # For bisection, only two initial values are needed
+  # # Inherited from bisection. Not used for now.
 
-  # ==== Set x_interval ====
+  # # This method only needs an initial interval
+  # if (inherits(object_by_org, "power4test_by_n") ||
+  #     inherits(object_by_org, "power4test_by_es")) {
+  #   x_i <- set_x_range_by_x(
+  #                     object,
+  #                     x = x,
+  #                     pop_es_name = pop_es_name,
+  #                     target_power = target_power,
+  #                     k = 2,
+  #                     x_max = x_max,
+  #                     x_min = x_min,
+  #                     object_by_org = object_by_org,
+  #                     what = what,
+  #                     goal = goal,
+  #                     tol = tol,
+  #                     ci_level = ci_level)
+  # } else {
+  #   x_i <- set_x_range(object,
+  #                     x = x,
+  #                     pop_es_name = pop_es_name,
+  #                     target_power = target_power,
+  #                     k = 2,
+  #                     x_max = x_max,
+  #                     x_min = x_min)
+  # }
 
-  if (x_include_interval) {
-    # For bisection, should exclude them initially,
-    # and let extend_interval() to do the job
-    x_i <- sort(c(x_interval, x_i))
-  }
-  x_i <- sort(unique(x_i))
-  if (length(x_i) == 1) {
-    x_interval_min <- min(x_interval)
-    x_interval_max <- max(x_interval)
-    if ((x_i < x_interval_min) ||
-        (x_i > x_interval_max)) {
-      x_i <- x_interval
-    } else {
-      tmp <- mean(x_interval)
-      if (x_i < tmp) {
-        x_i <- c(x_i, x_interval_max)
-      } else if (x_i > tmp) {
-        x_i <- c(x_i, x_interval_min)
-      } else {
-        x_i <- c(mean(c(x_i, x_interval_min)),
-                 mean(c(x_i, x_interval_max)))
-        x_i <- range(x_i)
-      }
-    }
-  } else {
-    x_i <- range(x_i)
-  }
+  # # For bisection, no need to exclude the value in the input objects
+  # # Exclude the value in the input object
+
+  # # ==== Set x_interval ====
+
+  # if (x_include_interval) {
+  #   # For bisection, should exclude them initially,
+  #   # and let extend_interval() to do the job
+  #   x_i <- sort(c(x_interval, x_i))
+  # }
+  # x_i <- sort(unique(x_i))
+  # if (length(x_i) == 1) {
+  #   x_interval_min <- min(x_interval)
+  #   x_interval_max <- max(x_interval)
+  #   if ((x_i < x_interval_min) ||
+  #       (x_i > x_interval_max)) {
+  #     x_i <- x_interval
+  #   } else {
+  #     tmp <- mean(x_interval)
+  #     if (x_i < tmp) {
+  #       x_i <- c(x_i, x_interval_max)
+  #     } else if (x_i > tmp) {
+  #       x_i <- c(x_i, x_interval_min)
+  #     } else {
+  #       x_i <- c(mean(c(x_i, x_interval_min)),
+  #                mean(c(x_i, x_interval_max)))
+  #       x_i <- range(x_i)
+  #     }
+  #   }
+  # } else {
+  #   x_i <- range(x_i)
+  # }
 
   # ==== Update by_x (by_x_i) ====