+++ NEWS.md
@@ -1,4 +1,4 @@
-# power4mome 0.1.1.55
+# power4mome 0.1.1.56
 
 - Improve the function for extending the
   initial interval before doing a
@@ -319,6 +319,29 @@
 - Updated `power4test()` to make one
   cluster that will be used in all stages.
   (0.1.1.55)
+
+- Load balancing is no longer used by
+  default, to ensure the results are
+  reproducible. To enable load balancing,
+  set the option `"power4mome.use_lb"`
+  to `TRUE` using `options()`.
+  (0.1.1.56)
+
+- For probabilistic bisection, the
+  initial interval will no longer be
+  adjusted. This algorithm should e used
+  with a wide enough initial interval
+  because the interval will not be
+  adjusted during the search (for now).
+  (0.1.1.56)
+
+- Updated the quick functions
+  (`q_power_mediation()` and friends)
+  to have one more mode, `"n"`.
+  Probabilistic bisection is the default
+  algorithm for this mode. Other methods
+  have been updated for this mode.
+  (0.1.1.56)
 
 # power4mome 0.1.1